+++ simulations/geometryXYVxVy/landau/pdi_out.yml.hpp
@@ -3,10 +3,16 @@
 
 constexpr char const* const PDI_CFG = R"PDI_CFG(
 metadata:
+)PDI_CFG"
+#ifdef SPLINE
+                                      R"PDI_CFG(
   Nx_spline_cells : int
   Ny_spline_cells : int
   Nvx_spline_cells : int
   Nvy_spline_cells : int
+)PDI_CFG"
+#endif
+                                      R"PDI_CFG(
   iter : int
   time_saved : double
   nbstep_diag: int
@@ -80,7 +86,17 @@ plugins:
     - file: 'GYSELALIBXX_initstate.h5'
       on_event: [initial_state]
       collision_policy: replace_and_warn
+)PDI_CFG"
+#ifdef SPLINE
+                                      R"PDI_CFG(
       write: [Nx_spline_cells, Nvx_spline_cells, MeshX, MeshY, MeshVx, MeshVy, nbstep_diag, Nkinspecies, fdistribu_charges, fdistribu_masses, fdistribu_eq]
+)PDI_CFG"
+#else
+                                      R"PDI_CFG(
+      write: [MeshX, MeshY, MeshVx, MeshVy, nbstep_diag, Nkinspecies, fdistribu_charges, fdistribu_masses, fdistribu_eq]
+)PDI_CFG"
+#endif
+                                      R"PDI_CFG(
     - file: 'GYSELALIBXX_${iter_saved:05}.h5'
       communicator: $MPI_COMM_WORLD
       on_event: [iteration, last_iteration]